TestableViewModel
