+++ PySubtrans/Helpers/TestCases.py
@@ -209,6 +209,48 @@ def BuildSubtitlesFromLineCounts(line_counts : list[list[int]]) -> Subtitles:
     subtitles.scenes = scenes
     return subtitles
 
+def CreateDummyBatch(scene_number : int, batch_number : int, line_count : int, start_line_number : int, start_time : timedelta) -> SubtitleBatch:
+    """
+    Helper to create a SubtitleBatch with the specified number of lines.
+    """
+    lines = [
+        SubtitleLine.Construct(
+            start_line_number + i,
+            start_time + timedelta(seconds=i*2),
+            start_time + timedelta(seconds=i*2 + 1),
+            f"Scene {scene_number} Batch {batch_number} Line {start_line_number + i}",
+            {}
+        )
+        for i in range(line_count)
+    ]
+
+    return SubtitleBatch({
+        'scene': scene_number,
+        'number': batch_number,
+        'summary': f"Scene {scene_number} Batch {batch_number}",
+        'originals': lines
+    })
+
+def CreateDummyScene(scene_number : int, batch_line_counts : list[int], start_line_number : int, start_time : timedelta) -> SubtitleScene:
+    """
+    Helper to create a SubtitleScene with batches containing the specified line counts.
+    """
+    batches = []
+    line_number = start_line_number
+    current_time = start_time
+
+    for batch_index, line_count in enumerate(batch_line_counts, start=1):
+        batch = CreateDummyBatch(scene_number, batch_index, line_count, line_number, current_time)
+        batches.append(batch)
+        line_number += line_count
+        current_time += timedelta(seconds=line_count * 2)
+
+    return SubtitleScene({
+        'number': scene_number,
+        'context': {'summary': f"Scene {scene_number}"},
+        'batches': batches
+    })
+
 class DummyProvider(TranslationProvider):
     name = "Dummy Provider"
